==> 1.Data Loading.ipynb <==
import pandas as pd
import json

with open('python_prg.JSON', 'r', encoding = "utf-8") as f:
    data = json.load(f)

all_comments = []

for repo, comments in data.items():
    if isinstance(comments, list):
        for c in comments:
            c["repo"] = repo   # add repo name
            all_comments.append(c)

df = pd.DataFrame(all_comments)

print(df.head())
print(df.columns)
print(len(df))


df.columns

df1 = df[['body', 'diff_hunk','path', 'repo']]
df1.rename(columns={'body': 'review_comments'}, inplace = True)

df1.isnull().any()

df1 = df1[df1['review_comments'].str.len()>20]
df1.shape

df1.to_feather('python_comm.feather')
print('done')

# df1.to_csv('python_comm.csv')
# print('done')
