==> 4.train_val_test_split.ipynb <==
import pandas as pd
import json

path = "train.jsonl"
data = []

with open(path, "r", encoding="utf-8") as f:
    for line in f:
        if line.strip():  # skip empty lines
            data.append(json.loads(line))

print("Number of samples:", len(data))

print(data[0])

from sklearn.model_selection import train_test_split

train_frac, val_frac, test_frac = 0.8, 0.1, 0.1

train_data, temp_data = train_test_split(data,train_size = train_frac, random_state = 42, shuffle = True)

rel_val = val_frac/(val_frac+test_frac)
val_data, test_data = train_test_split(temp_data, train_size = val_frac, random_state = 42, shuffle = True)
len(train_data), len(val_data), len(test_data)

with open("train_data.jsonl","w",encoding="utf-8") as f: 
    [f.write(json.dumps(x, ensure_ascii=False)+"\n") for x in train_data]
    
with open("val_data.jsonl","w",encoding="utf-8") as f: 
    [f.write(json.dumps(x, ensure_ascii=False)+"\n") for x in val_data]
    
with open("test_data.jsonl","w",encoding="utf-8") as f: 
    [f.write(json.dumps(x, ensure_ascii=False)+"\n") for x in test_data]
